Text changes for clarity
The ‘Show results for a genomic region’ card on the homepage matches the ‘Results’ for chr11: 33,858,576 - 33,892,076 (region) text shown on the search results page. Describing it as ‘Top significant effect observation’ might be more descriptive of what we are showing the user. Additionally, using ‘Filter Region’ could bring more clarity. Our goal is to make the portal as intuitive as possible.