Merge exons in genoverse
In the browser, when a gene is shown at "see exons" zoom level
all returned exons were being used to draw the gene. This was causing
extra lines to be shown because there were a lot "different" exons that
covered the same locations.

These changes update the code to merge all overlapping exons for a gene into
a single exon. The resulting gene drawing looks much better.