Support new fetch results format
Psycopg 3 now returns an actual list when the result is an array.
Before it would return a string that needed to be fully parsed.

Because the list used to be a string the result could be shoved in a set
for de-duplication. Now that it's an actual list it can't so we have to
change the method of deduplication.

Also, now only the location needs to be parsed, not the whole list.