Remove proximal REO queries from dna feature page
The proximal REO table is loaded async via htmx. But
we were still doing a count and actually loading the reos.

This uses the .exists method to avoid actually loading all the
proximal REOs and removes the pagination code that performs
the count.

These problems were especially noticable on MYB, which has a lot of
REOs.