Union facets on experiment cov page
Previously the facets were getting intersected, which could make it
impossible to filter on certain facet values. For instance, if one
experiment had "depleted", "enriched", and "non-sig" REOs and another
only had enriched, then the only "Direction" facet value shown would
be "enriched". I think union-ing solves the problem and is appropriate for
the cases when the experiments are both interesected and unioned together.