Merge cell line and tissue type into biosample
Merges the Cell Line and Tissue Type facets into a single
Biosample facet.

Previously it was assumed that each faceted object would
only have one value per facet, but now some items have two --
a cell line and a separate tissue type. This exposed a bug
in the SQL used to filter experiments/analyses so that SQL
had to be updated.