Major refactor of view code
Separated filter, stats, and bucket logic. Reos are only
iterated through once, filtering as needed. The reos that
aren't filtered out get passed to stats and chromosome
objects to generate stats and chromosome buckets, respectively.

This deduplicates a ton of code and makes the logic much, much easier
to follow.